+++ tests/cfgs.rs
@@ -1,4 +1,4 @@
-use pin_init::{pin_data, pin_init, PinInit};
+use pin_init::{pin_data, pin_init, stack_pin_init, PinInit};
 
 #[pin_data]
 pub struct Struct {
@@ -26,4 +26,27 @@ pub struct Struct2 {
     // Test for cases where the type is not even defined when cfg is not satisfied.
     #[cfg(any())]
     non_exist: NonExistentType,
+}
+
+#[pin_data]
+pub struct TupleStruct(#[cfg(any())] Field, u32, u32);
+
+impl TupleStruct {
+    pub fn new() -> impl PinInit<Self> {
+        pin_init!(Self {
+            #[cfg(any())]
+            0: Field,
+            // Disabled fields don't occupy an index!
+            0: 10,
+            1: 20,
+        })
+    }
+}
+
+#[test]
+fn tuple_fields_cfg_renumber() {
+    stack_pin_init!(let value = TupleStruct::new());
+    let proj = value.project();
+    assert_eq!(*proj.0, 10);
+    assert_eq!(*proj.1, 20);
 }