+++ tests/ui/compile-fail/init/tuple_missing_field.rs
@@ -1,0 +1,9 @@
+use pin_init::*;
+
+#[pin_data]
+struct Tuple(#[pin] i32, i32);
+
+fn main() {
+    let _ = pin_init!(Tuple { 0: 1 });
+    let _ = init!(Tuple { 0: 1 });
+}