+++ tests/ui/expand/tuple_struct.rs
@@ -4,4 +4,11 @@ use pin_init::*;
 #[pin_data]
 struct Foo<'a, T: Copy, const N: usize>(&'a mut [T; N], #[pin] PhantomPinned, usize);
 
-fn main() {}
+fn main() {
+    let mut first = [1u8, 2, 3];
+    let _ = init!(Foo {
+        0: &mut first,
+        1: PhantomPinned,
+        2 <- 10,
+    });
+}